+++ examples/scripts/ct_astra_modl_train_foam2.py
@@ -54,6 +54,11 @@ ray.init(logging_level=logging.ERROR)  # need to call init before jax import: ra
 
 import jax
 
+try:
+    from jax.extend.backend import get_backend  # introduced in jax 0.4.33
+except ImportError:
+    from jax.lib.xla_bridge import get_backend
+
 from mpl_toolkits.axes_grid1 import make_axes_locatable
 
 from scico import flax as sflax
@@ -67,7 +72,7 @@ Prepare parallel processing. Set an arbitrary processor count (only
 applies if GPU is not available).
 """
 os.environ["XLA_FLAGS"] = "--xla_force_host_platform_device_count=8"
-platform = jax.lib.xla_bridge.get_backend().platform
+platform = get_backend().platform
 print("Platform: ", platform)