+++ CHANGES.rst
@@ -20,7 +20,7 @@ Version 0.0.6   (unreleased)
 • Rename ``scico.flax.save_weights`` and ``scico.flax.load_weights`` to
   ``scico.flax.save_variables`` and ``scico.flax.load_variables``
   respectively.
-• Support ``jaxlib`` and ``jax`` versions 0.4.3 to 0.4.33.
+• Support ``jaxlib`` and ``jax`` versions 0.4.3 to 0.4.34.
 • Support ``flax`` versions 0.8.0 to 0.9.0.